Add value computation to frontier discovery node based on uncovered portions of the map
In this commit I've added the concept of using the uncovered area to compute frontier values.
Some authors prefer using more complex methods, however, in practice it is impossible to predict
what is to come. Consequently, authors, such as Burgard, mention that computing the area to be
uncovered given a sensor range is enough.